move protocol into server, reorg
