+++ src/server/renderer.ts
@@ -17,7 +17,7 @@ import {
   LoadFileInfo,
   RendererServerMessenger,
   RenderInfo,
-} from "../protocol/mod.ts";
+} from "./protocol/server-renderer.ts";
 import { Connection, createMessenger } from "../messages/mod.ts";
 import { registeredPlugins } from "./registerClientPlugin.ts";
 import { HashProduct } from "../core/mod.ts";
