+++ src/server/server.ts
@@ -7,12 +7,12 @@ import {
   EventEmitter,
   logConnection,
 } from "../messages/mod.ts";
+import { ServerClientMessenger } from "./protocol/server-client.ts";
 import {
-  ServerClientMessenger,
   ServerRendererMessenger,
   ServerRendererShape,
-} from "../protocol/mod.ts";
-import { ServerTranspilerMessenger } from "../protocol/mod.ts";
+} from "./protocol/server-renderer.ts";
+import { ServerTranspilerMessenger } from "./protocol/server-transpiler.ts";
 import { transformUrl } from "../server/transformUrl.ts";
 
 export interface ServerHost {
