+++ src/server/transpiler.ts
@@ -1,6 +1,6 @@
 import { ts } from "../deps/tsc.ts";
 import { stylus } from "../deps/stylus.ts";
-import { TranspilerServerMessenger } from "../protocol/mod.ts";
+import { TranspilerServerMessenger } from "./protocol/server-transpiler.ts";
 import { Connection, createMessenger } from "../messages/mod.ts";
 import { transformUrl } from "./transformUrl.ts";
